Add caller to token.send call
This commit adds an optional `caller` input argument to the token.send
call. By setting `caller` to the accountUpdate that's being used to
authorize a token transfer, we let the token contract be callable by
other zkApps.

If one zkApp calls on a token contract to authorize a transfer, the
`caller` field of the token contract AccountUpdate must be set to the
token id of the calling zkApp.