Prepend /bin/sh when running dump-test-vectors.sh
While this should be redundant, it is not in my experience.
Though I had the execution permission set, my system did not want to execute this script. It is not clear to me why, since the error would only say "file not found", hence I opted for this simple workaround.