feat(field.ts): add `and` method to `Field` class to perform bitwise AND operation on field elements
The `and` method is added to the `Field` class to perform bitwise AND operation on field elements. It takes two arguments, `y` and `length`. If both `this` and `y` are constants, it performs the operation using the `Fp.and` method. Otherwise, it uses the `Snarky.field.band` method to perform the operation. If the result of the operation is greater than `length` bits, an error is thrown.