docs(CHANGELOG.md): update unreleased changes and add version 0.12.0 details
This commit updates the unreleased changes link and adds the details of version 0.12.0. The breaking changes in version 0.12.0 are also documented, specifically the fix to the default verification key hash for AccountUpdates. This change adopts the default mechanism provided by Mina Protocol and may affect the verification key of already deployed contracts.