test(bitwise.unit-test.ts): add unit test for NOT operation with equivalentAsync
The unit test was added to test the NOT operation in the `bitwise.unit-test.ts` file. The test checks if the input value `x` is greater than or equal to 2^64 and throws an error if it does not fit into 64 bits. It then returns the result of the NOT operation on `x`. The test also includes an async function that uses the `Bitwise.not` method to calculate the NOT operation and returns the public output of the proof.