feat: add update-changelog.sh script to automate CHANGELOG.md updates
This script identifies the latest version number in the CHANGELOG, increments the patch segment of the version number for the new release, retrieves the current Git commit hash, and updates the CHANGELOG.md file accordingly. This change is done to automate the process of updating the CHANGELOG.md in response to new releases.