refactor(Int64): remove deprecated neg() method and rename negV2() to neg()
The deprecated `neg()` method has been removed from the `Int64` class. The `negV2()` method, which provides the correct implementation for negating an `Int64` value, has been renamed to `neg()`. This change ensures that the `neg()` method is backwards-compatible with future versions and follows the intended behavior.

docs(Int64): update method references and examples in JSDoc comments

The JSDoc comments have been updated to reflect the changes made to the `Int64` class. References to the deprecated `negV2()` method have been replaced with `neg()`. Similarly, references to `divV2()` and `modV2()` have been updated to `div()` and `mod()`, respectively. The corresponding examples in the JSDoc comments have also been updated to use the new method names.

refactor(Int64): remove TODO comment for isNegative() method

The TODO comment regarding the `isNegative()` method has been removed. The comment indicated that the method would be added when `checkV2` is enabled, but it seems that the method is already implemented correctly. Removing the comment helps to keep the code clean and avoids confusion.