test(account-update): uncomment and fix test case for AccountUpdate.check method
The test case for AccountUpdate.check method that checks if an error is thrown when both parentsOwnToken and inheritFromParent flags are true was commented out. This test case has been uncommented and fixed to ensure that the expected error is thrown in this scenario.

feat(random): add generator for MayUseToken type

A new generator for the MayUseToken type has been added to the random testing utilities. This generator creates random instances of the MayUseToken type with valid combinations of the parentsOwnToken and inheritFromParent flags. The generator is used to improve test coverage and ensure the correctness of the MayUseToken type.