refactor(run-ci-tests.sh): remove unit tests from CI test script
The unit tests have been removed from the CI test script. This change was likely made because the unit tests are now being run separately or have been moved to a different part of the CI pipeline. Removing them from this script helps to keep the CI process more focused and efficient.