js-layout-to-types: delegate prettier call to user
js-layout-to-types job should only be the generation of the files.
Prettifying the file should be done as a second step, by calling
`npm run format .`. This aligns with the UNIX philisophy that a script should be
doing only one thing, and well.

It removes some dependencies regarding prettier in the stanzas definition, and
simplify the addition of new plugins prettier might need.