Add a `Variant` class and to `lib/provable`
`Variant`s (alternatively known as "sum types",
"discriminated unions", or "tagged unions") are a
way to define a type that can be one of several
different types, but only one at a time.
This is useful for representing data that can
take on multiple forms, such as for JSON data
that can be one of several different types
(e.g. a string, number, object, array, etc.).